Add documentation + relocate services under a services folder
